Move intersect function to array-utils
