--- scv-prototype-frontend/src/app/components/PrivateRoute/index.jsx
@@ -1,1 +1,0 @@
-export { default } from './PrivateRoute';
